Fixed dao and service logic
