+++ pom.xml
@@ -50,6 +50,10 @@
 			<artifactId>hibernate-core</artifactId>
 			<version>5.6.3.Final</version>
 		</dependency>
+<!--		<dependency>-->
+<!--			<groupId>org.springframework.boot</groupId>-->
+<!--			<artifactId>spring-boot-starter-security</artifactId>-->
+<!--		</dependency>-->
 	</dependencies>
 
 	<build>
