+++ src/main/java/com/example/demo/models/exceptions/InvalidUsernameOrPasswordException.java
@@ -1,0 +1,4 @@
+package com.example.demo.models.exceptions;
+
+public class InvalidUsernameOrPasswordException extends RuntimeException{
+}
