+++ src/main/java/com/example/demo/models/exceptions/PasswordsDoNotMatchException.java
@@ -1,0 +1,8 @@
+package com.example.demo.models.exceptions;
+
+public class PasswordsDoNotMatchException extends RuntimeException{
+
+    public PasswordsDoNotMatchException() {
+        super("Passwords do not match exception.");
+    }
+}
