+++ src/main/java/com/example/demo/models/exceptions/PostalAddressAlreadyExistsException.java
@@ -1,9 +1,7 @@
 package com.example.demo.models.exceptions;
 
-import com.example.demo.models.PostalAddress;
-
 public class PostalAddressAlreadyExistsException extends RuntimeException{
-    public PostalAddressAlreadyExistsException(PostalAddress streetAddress) {
-        super(String.format("Postal address: %s already exists", streetAddress.toString()));
+    public PostalAddressAlreadyExistsException() {
+        super("Postal address: %s already exists.");
     }
 }
