+++ src/main/java/com/example/demo/models/exceptions/UsernameAlreadyExistsException.java
@@ -1,0 +1,8 @@
+package com.example.demo.models.exceptions;
+
+public class UsernameAlreadyExistsException extends RuntimeException{
+
+    public UsernameAlreadyExistsException(String username) {
+        super(String.format("User with username: %s already exists", username));
+    }
+}
