--- src/main/java/com/example/demo/repository/ContactMethodRepository.java
@@ -1,9 +1,0 @@
-package com.example.demo.repository;
-
-import com.example.demo.models.ContactMethod;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ContactMethodRepository extends JpaRepository<ContactMethod, Long> {
-}
